pyproject.toml
[tool.poetry]
name = "synthetic-data-generation"
version = "0.1.3"
description = "Algorithms for generating synthetic data"
authors = ["Daan Knoors <d.knoors@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
numpy = "^1.21.2"
pandas = "^1.3.2"
seaborn = "^0.11.2"
thomas-core = "^0.1.3"
pyjanitor = "^0.21.2"
diffprivlib = "^0.4.1"
dill = "^0.3.4"
lifelines = "^0.26.0"
dython = "^0.6.7"
scipy = "^1.7.1"
scikit-learn = "^0.24.2"
matplotlib = "^3.4.3"
joblib = "^1.0.1"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.1.10"
pandas-profiling = "^3.0.0"
PyQt5 = "^5.15.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
